requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SpeechRecognition==3.10.0
Werkzeug==2.3.7
openai-whisper==20231117
python-dotenv==1.0.0
opencc-python-reimplemented==0.1.7
numpy<2.3